wget: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@43256 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/Networking/Apps/wget/src/http.c
@@ -761,7 +761,7 @@ print_server_response (const struct response *resp, const char *prefix)
 	--e;
       /* This is safe even on printfs with broken handling of "%.<n>s"
 	 because resp->headers ends with \0.  */
-      logprintf (LOG_VERBOSE, "%s%.*s\n", prefix, e - b, b);
+      logprintf (LOG_VERBOSE, "%s%.*s\n", prefix, (int)(e - b), b);
     }
 }
